+++ packages/agent/src/agent.ts
@@ -19,6 +19,7 @@ import bodyParser from 'koa-bodyparser';
 import FrameworkMounter from './framework-mounter';
 import makeRoutes from './routes';
 import makeServices from './services';
+import ActionCustomizationService from './services/model-customizations/action-customization';
 import { AgentOptions, AgentOptionsWithDefaults } from './types';
 import SchemaGenerator from './utils/forest-schema/generator';
 import OptionsValidator from './utils/options-validator';
@@ -36,6 +37,8 @@ import OptionsValidator from './utils/options-validator';
 export default class Agent<S extends TSchema = TSchema> extends FrameworkMounter {
   private options: AgentOptionsWithDefaults;
   private customizer: DataSourceCustomizer<S>;
+  private nocodeCustomizer: DataSourceCustomizer<S>;
+  private actionCustomizationService: ActionCustomizationService;
 
   /**
    * Create a new Agent Builder.
@@ -59,26 +62,29 @@ export default class Agent<S extends TSchema = TSchema> extends FrameworkMounter
 
     this.options = allOptions;
     this.customizer = new DataSourceCustomizer<S>();
+    this.actionCustomizationService = new ActionCustomizationService(allOptions);
   }
 
   /**
    * Start the agent.
    */
   async start(): Promise<void> {
-    const { isProduction, logger, skipSchemaUpdate, typingsPath, typingsMaxDepth } = this.options;
+    const router = await this.buildRouterAndSendSchema();
 
-    const dataSource = await this.customizer.getDataSource(logger);
-    const [router] = await Promise.all([
-      this.getRouter(dataSource),
-      !skipSchemaUpdate ? this.sendSchema(dataSource) : Promise.resolve(),
-      !isProduction && typingsPath
-        ? this.customizer.updateTypesOnFileSystem(typingsPath, typingsMaxDepth)
-        : Promise.resolve(),
-
-      this.options.forestAdminClient.subscribeToServerEvents(),
-    ]);
+    await this.options.forestAdminClient.subscribeToServerEvents();
+    this.options.forestAdminClient.onRefreshCustomizations(this.restart.bind(this));
 
     await this.mount(router);
+  }
+
+  /**
+   * Restart the agent at runtime (remount routes).
+   */
+  private async restart(): Promise<void> {
+    // We force sending schema when restarting
+    const updatedRouter = await this.buildRouterAndSendSchema();
+
+    await this.remount(updatedRouter);
   }
 
   /**
@@ -163,23 +169,57 @@ export default class Agent<S extends TSchema = TSchema> extends FrameworkMounter
     return router;
   }
 
+  private async buildRouterAndSendSchema(): Promise<Router> {
+    const { isProduction, logger, typingsPath, typingsMaxDepth, experimental } = this.options;
+
+    // It allows to rebuild the full customization stack with no code customizations
+    this.nocodeCustomizer = new DataSourceCustomizer<S>();
+    this.nocodeCustomizer.addDataSource(this.customizer.getFactory());
+    this.nocodeCustomizer.use(
+      this.actionCustomizationService.addWebhookActions,
+      experimental?.webhookCustomActions,
+    );
+
+    const dataSource = await this.nocodeCustomizer.getDataSource(logger);
+    const [router] = await Promise.all([
+      this.getRouter(dataSource),
+      this.sendSchema(dataSource),
+      !isProduction && typingsPath
+        ? this.customizer.updateTypesOnFileSystem(typingsPath, typingsMaxDepth)
+        : Promise.resolve(),
+    ]);
+
+    return router;
+  }
+
   /**
    * Send the apimap to forest admin server
    */
   private async sendSchema(dataSource: DataSource): Promise<void> {
-    const { schemaPath } = this.options;
+    const { schemaPath, skipSchemaUpdate, isProduction, experimental } = this.options;
+
+    // skipSchemaUpdate is mainly used in cloud version
+    if (skipSchemaUpdate) {
+      this.options.logger(
+        'Warn',
+        'Schema update was skipped (caused by options.skipSchemaUpdate=true)',
+      );
+
+      return;
+    }
 
     // Either load the schema from the file system or build it
     let schema: ForestSchema;
 
-    if (this.options.isProduction) {
+    // When using experimental no-code features even in production we need to build a new schema
+    if (!experimental?.webhookCustomActions && isProduction) {
       try {
         schema = JSON.parse(await readFile(schemaPath, { encoding: 'utf-8' }));
       } catch (e) {
         throw new Error(`Can't load ${schemaPath}. Providing a schema is mandatory in production.`);
       }
     } else {
-      schema = await SchemaGenerator.buildSchema(dataSource);
+      schema = await SchemaGenerator.buildSchema(dataSource, this.buildSchemaFeatures());
 
       const pretty = stringify(schema, { maxLength: 100 });
       await writeFile(schemaPath, pretty, { encoding: 'utf-8' });
@@ -192,5 +232,15 @@ export default class Agent<S extends TSchema = TSchema> extends FrameworkMounter
       : 'Schema was not updated since last run';
 
     this.options.logger('Info', message);
+  }
+
+  private buildSchemaFeatures(): string[] | null {
+    const mapping: Record<keyof AgentOptions['experimental'], string> = {
+      webhookCustomActions: ActionCustomizationService.FEATURE,
+    };
+
+    return Object.entries(mapping)
+      .filter(([experimentalFeature]) => this.options.experimental?.[experimentalFeature])
+      .map(([, feature]) => feature);
   }
 }