+++ packages/forestadmin-client/src/build-application-services.ts
@@ -2,7 +2,10 @@ import AuthService from './auth';
 import ChartHandler from './charts/chart-handler';
 import EventsSubscriptionService from './events-subscription';
 import NativeRefreshEventsHandlerService from './events-subscription/native-refresh-events-handler-service';
+import { RefreshEventsHandlerService } from './events-subscription/types';
 import IpWhiteListService from './ip-whitelist';
+import ModelCustomizationFromApiService from './model-customizations/model-customization-from-api';
+import { ModelCustomizationService } from './model-customizations/types';
 import ActionPermissionService from './permissions/action-permission';
 import PermissionService from './permissions/permission-with-cache';
 import RenderingPermissionService from './permissions/rendering-permission';
@@ -28,7 +31,9 @@ export default function buildApplicationServices(
   permission: PermissionService;
   chartHandler: ChartHandler;
   auth: AuthService;
+  modelCustomizationService: ModelCustomizationService;
   eventsSubscription: EventsSubscriptionService;
+  eventsHandler: RefreshEventsHandlerService;
 } {
   const optionsWithDefaults = {
     forestServerUrl: 'https://api.forestadmin.com',
@@ -72,9 +77,14 @@ export default function buildApplicationServices(
     permission,
     contextVariables,
     eventsSubscription,
+    eventsHandler,
     chartHandler: new ChartHandler(contextVariables),
     ipWhitelist: new IpWhiteListService(optionsWithDefaults),
     schema: new SchemaService(optionsWithDefaults),
     auth: new AuthService(optionsWithDefaults),
+    modelCustomizationService: new ModelCustomizationFromApiService(
+      forestAdminServerInterface,
+      optionsWithDefaults,
+    ),
   };
 }