+++ packages/forestadmin-client/src/index.ts
@@ -42,7 +42,9 @@ export default function createForestAdminClient(
     ipWhitelist,
     schema,
     auth,
+    modelCustomizationService,
     eventsSubscription,
+    eventsHandler,
   } = buildApplicationServices(new ForestHttpApi(), options);
 
   return new ForestAdminClientWithCache(
@@ -54,12 +56,16 @@ export default function createForestAdminClient(
     ipWhitelist,
     schema,
     auth,
+    modelCustomizationService,
     eventsSubscription,
+    eventsHandler,
   );
 }
 
 export * from './charts/types';
 export * from './schema/types';
+export * from './model-customizations/types';
+
 export { default as ContextVariablesInjector } from './utils/context-variables-injector';
 export { default as ContextVariables } from './utils/context-variables';
 export { default as ChartHandler } from './charts/chart-handler';