+++ packages/forestadmin-client/test/__factories__/forest-admin-server-interface.ts
@@ -6,4 +6,5 @@ export default Factory.define<ForestAdminServerInterface>(() => ({
   getRenderingPermissions: jest.fn(),
   getEnvironmentPermissions: jest.fn(),
   getUsers: jest.fn(),
+  getModelCustomizations: jest.fn(),
 }));