+++ packages/datasource-customizer/CHANGELOG.md
@@ -1,3 +1,16 @@
+# @forestadmin/datasource-customizer [1.8.0-alpha-webhook-custom-actions.2](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-customizer@1.8.0-alpha-webhook-custom-actions.1...@forestadmin/datasource-customizer@1.8.0-alpha-webhook-custom-actions.2) (2023-06-07)
+
+
+### Bug Fixes
+
+* **typings:** allow incomplete records to be passed to collection.create ([#700](https://github.com/ForestAdmin/agent-nodejs/issues/700)) ([95483e2](https://github.com/ForestAdmin/agent-nodejs/commit/95483e277d695a9f3ed19d86b2f6834413007afc))
+
+# @forestadmin/datasource-customizer [1.8.0-alpha-webhook-custom-actions.1](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-customizer@1.7.1...@forestadmin/datasource-customizer@1.8.0-alpha-webhook-custom-actions.1) (2023-04-28)
+
+
+### Features
+
+* **webhook-custom-actions:** use webhook nocode smart actions configured in the frontend ([#671](https://github.com/ForestAdmin/agent-nodejs/issues/671)) ([7629699](https://github.com/ForestAdmin/agent-nodejs/commit/762969922c98ad4b15ee2407d5789ab939059020))
 ## @forestadmin/datasource-customizer [1.7.2](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-customizer@1.7.1...@forestadmin/datasource-customizer@1.7.2) (2023-05-24)
 
 
@@ -26,6 +39,13 @@
 ### Dependencies
 
 * **@forestadmin/datasource-toolkit:** upgraded to 1.5.0
+
+# @forestadmin/datasource-customizer [1.7.0-alpha-webhook-custom-actions.1](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-customizer@1.6.2...@forestadmin/datasource-customizer@1.7.0-alpha-webhook-custom-actions.1) (2023-04-20)
+
+
+### Features
+
+* **webhook-custom-actions:** use webhook nocode smart actions configured in the frontend ([#671](https://github.com/ForestAdmin/agent-nodejs/issues/671)) ([7629699](https://github.com/ForestAdmin/agent-nodejs/commit/762969922c98ad4b15ee2407d5789ab939059020))
 
 ## @forestadmin/datasource-customizer [1.6.2](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-customizer@1.6.1...@forestadmin/datasource-customizer@1.6.2) (2023-04-19)