+++ packages/plugin-aws-s3/CHANGELOG.md
@@ -1,3 +1,13 @@
+## @forestadmin/plugin-aws-s3 [1.2.1-alpha-webhook-custom-actions.1](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-aws-s3@1.2.0...@forestadmin/plugin-aws-s3@1.2.1-alpha-webhook-custom-actions.1) (2023-06-07)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.8.0-alpha-webhook-custom-actions.2
+
 # @forestadmin/plugin-aws-s3 [1.2.0](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-aws-s3@1.1.5...@forestadmin/plugin-aws-s3@1.2.0) (2023-06-02)
 
 
@@ -35,6 +45,16 @@
 
 * **@forestadmin/datasource-customizer:** upgraded to 1.7.0
 * **@forestadmin/datasource-toolkit:** upgraded to 1.5.0
+
+## @forestadmin/plugin-aws-s3 [1.1.3-alpha-webhook-custom-actions.1](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-aws-s3@1.1.2...@forestadmin/plugin-aws-s3@1.1.3-alpha-webhook-custom-actions.1) (2023-04-20)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.7.0-alpha-webhook-custom-actions.1
 
 ## @forestadmin/plugin-aws-s3 [1.1.2](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-aws-s3@1.1.1...@forestadmin/plugin-aws-s3@1.1.2) (2023-04-19)